+++ apps/save-editor/lib/l10n/app_localizations_fr.dart
@@ -2114,6 +2114,9 @@ class AppLocalizationsFr extends AppLocalizations {
   String get fallbackItem => 'Objet';
 
   @override
+  String get itemNaturalWeapon => 'Arme naturelle';
+
+  @override
   String get attributeSkillPointsFallback => 'Points d’apprentissage (PA)';
 
   @override