+++ apps/save-editor/lib/l10n/app_localizations_ru.dart
@@ -2112,6 +2112,9 @@ class AppLocalizationsRu extends AppLocalizations {
   String get fallbackItem => 'Предмет';
 
   @override
+  String get itemNaturalWeapon => 'Природное оружие';
+
+  @override
   String get attributeSkillPointsFallback => 'Очки обучения (LP)';
 
   @override